+++ src/main/java/tech/ignitr/archive/model/DemoData.java
@@ -1,4 +1,4 @@
-package tech.ignitr.character;
+package tech.ignitr.archive.model;
 
 /**
  *
