+++ src/main/java/tech/ignitr/archive/model/SuperZombie.java
@@ -1,4 +1,4 @@
-package tech.ignitr.character;
+package tech.ignitr.archive.model;
 
 import java.util.Objects;
